Fix missing QP in the activate modal
